--- .idea/jsLinters/eslint.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="EslintConfiguration">
-    <custom-configuration-file used="true" path="W:\ts-app\.eslintrc" />
-    <option name="fix-on-save" value="true" />
-  </component>
-</project>
