+++ prettier.config.js
@@ -1,9 +1,9 @@
 module.exports = {
-  printWidth: 80,
+  printWidth: 10,
   singleQuote: true,
   trailingComma: 'all',
   bracketSpacing: true,
   jsxBracketSameLine: false,
   tabWidth: 2,
-//  semi: false,
+  semi: false,
 }
