+++ src/app/App.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import logo from '../img/logo.svg';
+import './App.css';
+import Header from '../components/Header/Header';
+
+function App() {
+  return (
+    <div className="App">
+           <Header />
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+      </header>
+    </div>
+  );
+}
+
+export default App;
