+++ src/app/mockaData.tsx
@@ -1,0 +1,6 @@
+const mockData = {
+  str: 'new string',
+  numbers: [1, 2, 3, 4, 5, 6, 7],
+};
+
+export default mockData;
