+++ src/app/store.ts
@@ -1,11 +1,9 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import counterReducer from '../features/counter/counterSlice';
-import bookReducer from '../features/book/bookSlice';
 
 export const store = configureStore({
   reducer: {
-    counter: counterReducer,
-    book: bookReducer,
+    // counter: counterReducer,
+    // book: bookReducer,
   },
 });
