+++ src/components/Header/Header.css
@@ -1,0 +1,6 @@
+.header {
+  text-align: center;
+  display: grid;
+  place-items: center;
+}
+
