+++ src/components/Header/Header.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import './Header.css';
+
+function Header() {
+  return (
+          <div className="header">
+      Header
+    </div>
+  );
+}
+export default Header;
