--- src/components/diagram/Diagram.module.css
@@ -1,13 +1,0 @@
-.diagram {
-  width: 500px;
-  height: 400px;
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  margin-bottom: 48px;
-}
-
-.column {
-  width: 66px;
-  background-color: royalblue;
-}
