--- src/models/Person.ts
@@ -1,5 +1,0 @@
-export default interface Person {
-  name: string;
-  age: number;
-  role: 'admin' | 'user' | 36 | [];
-}
