+++ src/other/react-app-env.d.ts
